// src/app/Models/FormModel.ts
export class FormPostModel {
    Field8: string;
    Field9: string;
    Field10: string;
    Field11: string;
    Field12: string;
    Field13: string;
    Field14: string;
    Field15: string;
    Field16: string;
    Field17: string;
    Field18: string;
    Field19: string;
    Field20: string;
    Field21: string;
    Field22: string;
    Field23: string;
    Field24: string;
    Field25: string;
}